+++ lib/sidekiq/failures/web_extension.rb
@@ -23,7 +23,7 @@ module Sidekiq
           Sidekiq.redis {|c|
             c.multi do
               c.del("failed")
-              c.del("stat:failed") if params["counter"]
+              c.set("stat:failed", 0) if params["counter"]
             end
           }